+++ TestProjects/UniversalGraphicsTest_Foundation/Assets/Scenes/210_Motion_Vectors/CaptureMotionVectorsPass.cs
@@ -1,0 +1,51 @@
+using UnityEngine;
+using UnityEngine.Rendering;
+using UnityEngine.Rendering.Universal;
+
+internal class CaptureMotionVectorsPass : ScriptableRenderPass
+{
+    ProfilingSampler m_ProfilingSampler = new ProfilingSampler("MotionVecDebug");
+    Material m_Material;
+    RenderTargetIdentifier m_CameraColorTarget;
+    float m_intensity;
+
+    public CaptureMotionVectorsPass(Material material)
+    {
+        m_Material = material;
+        renderPassEvent = RenderPassEvent.BeforeRenderingPostProcessing;
+    }
+
+    public void SetTarget(RenderTargetIdentifier colorHandle, float intensity)
+    {
+        m_CameraColorTarget = new RenderTargetIdentifier(colorHandle, 0, CubemapFace.Unknown, -1);
+        m_intensity = intensity;
+    }
+
+    public override void Configure(CommandBuffer cmd, RenderTextureDescriptor cameraTextureDescriptor)
+    {
+        ConfigureTarget(m_CameraColorTarget, m_CameraColorTarget);
+        ConfigureClear(ClearFlag.Color, Color.black);
+    }
+
+    public override void Execute(ScriptableRenderContext context, ref RenderingData renderingData)
+    {
+        var camera = renderingData.cameraData.camera;
+        if (camera.cameraType != CameraType.Game)
+            return;
+
+        if (m_Material == null)
+            return;
+
+        //Todo: test code is not working for XR
+        CommandBuffer cmd = CommandBufferPool.Get();
+        using (new ProfilingScope(cmd, m_ProfilingSampler))
+        {
+            m_Material.SetFloat("_Intensity", m_intensity);
+            cmd.DrawMesh(RenderingUtils.fullscreenMesh, Matrix4x4.identity, m_Material);
+        }
+        context.ExecuteCommandBuffer(cmd);
+        cmd.Clear();
+
+        CommandBufferPool.Release(cmd);
+    }
+}